tinking with test_translator.py
